requirement.txt
# Python packages requirements
Pillow~=9.2.0
pyserial~=3.5
PyYAML~=6.0
psutil~=5.9.1
GPUtil~=1.4.0
# Following packages are for AMD GPU on Linux only
Cython~=0.29.32; os_name=='posix'
pyamdgpuinfo~=2.1.3; os_name=='posix'